Extract DEM grid boundary alignment policy
